+++ VStore.Host/ActionResults/ConflictResult.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Mvc;
+
+namespace NuClear.VStore.Host.ActionResults
+{
+    public sealed class ConflictResult : StatusCodeResult
+    {
+        public ConflictResult() : base(409)
+        {
+        }
+    }
+}
