--- VStore.Host/Controllers/LockSessionManagementController.cs
@@ -1,32 +1,0 @@
-﻿using System.Threading.Tasks;
-
-using Microsoft.AspNetCore.Mvc;
-
-using NuClear.VStore.Locks;
-
-namespace NuClear.VStore.Host.Controllers
-{
-    [Route("mgmt/lock-session")]
-    public sealed class LockSessionManagementController : Controller
-    {
-        private readonly LockSessionManager _lockSessionManager;
-
-        public LockSessionManagementController(LockSessionManager lockSessionManager)
-        {
-            _lockSessionManager = lockSessionManager;
-        }
-
-        [HttpGet]
-        public async Task<JsonResult> ListAllCurrentLockSessios()
-        {
-            var keys = await _lockSessionManager.GetAllCurrentLockSessions();
-            return Json(keys);
-        }
-
-        [HttpDelete("{rootObjectId}")]
-        public async Task DeleteSessionLock(string rootObjectId)
-        {
-            await _lockSessionManager.DeleteLockSession(rootObjectId);
-        }
-    }
-}
