--- VStore.Host/Convensions/MvcOptionsExtensions.cs
@@ -1,13 +1,0 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Routing;
-
-namespace NuClear.VStore.Host.Convensions
-{
-    public static class MvcOptionsExtensions
-    {
-        public static void UseGlobalRoutePrefix(this MvcOptions opts, IRouteTemplateProvider routeAttribute)
-        {
-            opts.Conventions.Insert(0, new RouteConvention(routeAttribute));
-        }
-    }
-}
