+++ VStore.Host/project.json
@@ -5,10 +5,9 @@
     "AWSSDK.Extensions.NETCore.Setup": "3.3.0.2",
     "Microsoft.NETCore.Runtime.CoreCLR": "1.1.0",
     "Microsoft.NETCore.DotNetHostPolicy": "1.1.0",
-    "Microsoft.AspNetCore.Cors": "1.1.0",
     "Microsoft.AspNetCore.Diagnostics": "1.1.0",
     "Microsoft.AspNetCore.Mvc": "1.1.0",
-    "Microsoft.AspNetCore.Routing": "1.1.0",
+    "Microsoft.AspNetCore.Mvc.Versioning": "1.0.3",
     "Microsoft.AspNetCore.Server.Kestrel": "1.1.0",
     "Microsoft.Extensions.Configuration.EnvironmentVariables": "1.1.0",
     "Microsoft.Extensions.Configuration.FileExtensions": "1.1.0",
